app/requirements.txt
Django==2.2.3
djangorestframework==3.10.0
pytz==2019.1
pika==1.0.1
requests==2.22.0
